requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv==1.0.1
discord.py==2.4.0
